users_service/pyproject.toml
[project]
name = "users-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "pydantic>=2.11.9",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
]
